break: improve navbar link behavior using components
